requirements.txt
bitarray==2.9.2
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
DateTime==5.5
decorator==5.2.1
env==0.1.0
et_xmlfile==2.0.0
Flask==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
knot==0.4.0
knot_floer_homology==1.2.2
lxml==6.0.0
MarkupSafe==3.0.2
networkx==3.5
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pdfminer==20191125
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==10.2.0
pycparser==2.22
pycryptodome==3.20.0
PyMuPDF==1.26.3
pypdfium2==4.30.1
python-dateutil==2.8.2
python-docx==1.2.0
pytz==2025.2
setuptools==80.9.0
six==1.16.0
snappy_manifolds==1.2.1
spherogram==2.3
SQLite3-0611==0.0.1
SQLite4==0.1.1
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
xlrd==2.0.2
zope.interface==7.2
